hide backend solver declarations from public include
