+++ lib/Solver/Z3Solver.h
@@ -1,0 +1,34 @@
+//===-- Z3Solver.h
+//---------------------------------------------------===//
+//
+//                     The KLEE Symbolic Virtual Machine
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef KLEE_Z3SOLVER_H
+#define KLEE_Z3SOLVER_H
+
+#include "klee/Solver.h"
+
+namespace klee {
+/// Z3Solver - A solver complete solver based on Z3
+class Z3Solver : public Solver {
+public:
+  /// Z3Solver - Construct a new Z3Solver.
+  Z3Solver();
+
+  /// Get the query in SMT-LIBv2 format.
+  /// \return A C-style string. The caller is responsible for freeing this.
+  virtual char *getConstraintLog(const Query &);
+
+  /// setCoreSolverTimeout - Set constraint solver timeout delay to the given
+  /// value; 0
+  /// is off.
+  virtual void setCoreSolverTimeout(double timeout);
+};
+}
+
+#endif /* KLEE_Z3SOLVER_H */
